+++ scripts/lib/backfill-utils.ts
@@ -1,17 +1,12 @@
 /** Shared helpers for backfill-2025* scripts. */
 
-import { prisma } from "../../server/src/lib/prisma";
 import { MIN_TERM_LENGTH } from "../../utils/constants";
+import { prisma } from "./d1-prisma";
 
-/**
- * Count properties scraped for year 2025.
- *
- * Uses ::int cast because Prisma $queryRaw returns BigInt for COUNT(*);
- * without the cast the `number` annotation silently lies (BigInt !== number).
- */
+/** Count properties scraped for year 2025. */
 export async function get2025Count(): Promise<number> {
 	const result = await prisma.$queryRaw<[{ count: number }]>`
-    SELECT COUNT(*)::int as count FROM properties WHERE year = 2025`;
+    SELECT COUNT(*) as count FROM properties WHERE year = 2025`;
 	return result[0].count;
 }
